Implement issue status updating for JIRA backend
Implement issue status updates for the JIRA backend. Due to some
annoying (and not really documented) restrictions of the JIRA API, this
has to happen as a *separate* API call (to "transitions") from actually
updating the fields of the issue, and we need to do some extra work to
look up the ID of the state transition.