Try to re-populate a null epic field
If the epic-custom-field is bound but null, still try to look it up from
the API (don't treat it as cached). I saw this happen once and I'm not
entirely sure why (perhaps an error condition that got it into a bad
state) but generally I'd prefer this work but be slow right now than
break permanently.