Don't crash if completion-metadata is called in ivy-read
If icomplete-mode is active, a hook seems to be installed that causes
ivy-read to call completion-metadata on our lambda. This function passes
an extra of two arguments to its table argument (our lambda) to encode
the request for metadata.

We can simply ignore these extra arguments since completion-metadata
will interpret a non-list return value as “no metadata available”.